+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # New
 - Increase MSRV to 1.50.0
 - Fix `Number::as_[u128,i128]` unsoundness issues (#29)
+- Fix number parsing, preferring integer parsing (#33)
 
 # 0.4.0
 - Add `Borrow` trait to `Kstr`